Initial commit of sweetbasil.rocks rewrite
This project builds on what I learned in my initial project https://github.com/iancullinane/sweetbasil.rocks

This rewrite integrates reactjs and removes poorly designed and redundant code. I feel like I changed enough that I wanted to make a new project from it. It isn't a full react app but the track list on the main page is built with reactjs. This project also uses babel in a few places, and has make provisions for watching a script. Tested with browsersync but not currently under version control.

Author:  Ian Cullinane <ian.cullinane@gmail.com>